ISSUE-17 and ISSUE-25: Remove the KeyAttributes object
In light of the Key object following the structured clone, the following is true:
- For Keys created by an application (KeyGenerator), all attributes can be stored
  in existing web storage mechanisms. The need for 'read-only' attributes is based
  on a presumption of defense against XSS - which is not the intent. As such,
  applications can associate whatever attributes they want with a Key, in the same
  storage mechanism of the Key itself, and simply note not to modify the attributes.
- For Keys that are provided by an "external" source (eg: pre-provisioned device
  specific keys), the attributes - and means of exposing these attributes - are
  defined by the source itself. Since these attributes are either implementation
  or application specific, they should be dictated by the spec that details how
  the implementation/application behaves.

Applications that wish to pass around Key objects with custom properties can
always do so through extending or wrapping the object, obtaining and/or enumerating
the properties from the underlying key storage backend. By NOT specifying this
behaviour, it allows maximum flexibility for both application developers and
implementors, and avoids having to try to normalize all possible naming and
attribute schemes into a single API pattern.