Rename Crypto -> SubtleCrypto, placing it in window.crypto.subtle
One of the ongoing points of feedback/confusion is regarding the exposure of
low-level algorithms to web developers. Some view this as an inherently
dangerous proposition, because web developers can't or don't understand
cryptographic tradeoffs. However, such a low-level interface is necessary to
support a variety of web applications, including applications or libraries that
provide secure high-level compositions.

Because protocols vary in requirements, and because these compositions vary in
implementation, a low-level API is necessary to support them. However, to make
it clearer to developers that this is subtle and easy to get wrong (just like
any other aspect of development, if not more so), move the existing Crypto
interface into SubtleCrypto.