Bug 23728, Bug 24878, Bug 25436 - Formally specify the algorithm normalization process.
This is accomplished through reviving the concept of 'registered algorithms', for the
purpose of explicitly defining how conversion from ECMAScript objects to IDL types
takes place.

This allows individual algorithms to eliminate common boilerplate, as well as make
sure that handling for (optional fields, CryptoOperationData) are consistently
adhered to.

It's a step further into treating algorithms as "internal objects" with state, which
is to allow for greater extensibility (by describing lookups in internal maps),
so that new algorithms do not need to update the 'main' specification.

In the course of this, two changes were made, neither of which fundamentally alters
the API in a way that users can distinguish - the introduction of a typedef called
HashAlgorithmIdentifier that is an AlgorithmIdentifier, and simply used for specifying
how nested hash algorithms work (which are a sort of special case), and the restructuring
of the algorithms section to better explain and explore these concepts.