Bug 25094: Add a length parameter to the HmacKeyParams. This also aligns the HmacImportParams with the HmacKeyGenParams
BREAKING CHANGE: Implementations that have already allowed HMAC keys to be imported will not have recorded the length. The length
of the keys can be derived from imported keys, since only byte-length multiples are supported. However, for keys that were
generated, if the length was not byte-aligned, this can create backwards-compat issues.