Check for a wrap of the counter
Need to ensure that the counter does not wrap when doing an encrypt
because otherwise there is a bad leakage of information.

We cannot do the full check on this - would need to also have a check
for starting at zero or something so just do the best we can.  The
algorithm for counters referenced permits the counter to wrap mod 2^m