Remove dead writes to CryptoKey.extractable in HKDF/PBKDF2.importKey
When these steps execute, it is know due to an earlier step that the
argument 'extractable' is false, and that the internal slot will not be
read before being overwritten again by SubtleCrypto.importKey:

https://w3c.github.io/webcrypto/#SubtleCrypto-method-importKey

Therefore, it is a dead write. Removing it makes it more similar to all
other importKey operations, which do not write the [[extractable]]
internal slot either.